fix types, versioning, etc
