+++ packages/core/src/sharedTypes/index.ts
@@ -1,3 +1,1 @@
 export * from "./eventLog.js"
-export * from "./mentalProcess.js"
-export type { CortexStep } from "./cortexStep.d.ts"
