+++ packages/soul/src/event-log.ts
@@ -1,6 +1,6 @@
 import { syncedStore } from "@syncedstore/core"
 
-import { eventLogShape, type EventLogDoc, type MentalProcess } from "@opensouls/core"
+import { eventLogShape, type EventLogDoc } from "@opensouls/core"
 
 export const syncedEventStore = 
   (): ReturnType<typeof syncedStore<EventLogDoc>> => syncedStore<EventLogDoc>(eventLogShape)
